Crate vivcord

Re-exports

pub use api::ApiClient;

pub use gateway::Gateway;

pub use datatypes::Intents;

Modules

api

Code for interacting with the discord REST api

datatypes

Types used by discord

gateway

Connect and recive events from discord

Macros

to_snowflake_simple

Implement From<...> for Snowflake This should be used on structs where there is a id field

wait_for

Wait for specific event from gateway

Structs

CreateMessageParams

Fields that can be passed to the discord api to create message This could be the create_message endpoint, or TODO: MORE ENDPOINTS

Enums

GatewayEventData

Discord Events, returned from Gateway